asynchronously call setState() in a safe way
In order to safely call setState() asynchronously, the "mounted" property of the State object (in this case, "_IntlPhoneFieldState"), needs to be validated. If mounted is not true, setState should not be called because it would result in the exception "Unhandled Exception: setState() called after dispose(), but did not call dispose method", since the widget was disposed and its state is no longer mounted.